Modularize JavaScript code
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/julienbro/WallSim3d?shareId=XXXX-XXXX-XXXX-XXXX).
