+++ src/domElements.js
@@ -1,0 +1,41 @@
+export const viewportContainer = document.getElementById('viewport-container');
+export const toolbar = document.getElementById('toolbar');
+export const elementSelect = document.getElementById('element-select');
+export const jointInput = document.getElementById('joint-thickness');
+export const infoText = document.querySelector('.info-text');
+export const selectToolButton = document.getElementById('tool-select');
+export const addToolButton = document.getElementById('tool-add');
+export const moveToolButton = document.getElementById('tool-move');
+export const rotateToolButton = document.getElementById('tool-rotate');
+export const duplicateToolButton = document.getElementById('tool-duplicate');
+export const deleteToolButton = document.getElementById('tool-delete');
+export const toolButtons = [selectToolButton, addToolButton, moveToolButton, rotateToolButton, duplicateToolButton, deleteToolButton];
+export const customElementGroup = document.getElementById('custom-element-group');
+export const customNameInput = document.getElementById('custom-name');
+export const customWidthInput = document.getElementById('custom-width');
+export const customHeightInput = document.getElementById('custom-height');
+export const customDepthInput = document.getElementById('custom-depth');
+export const btnOrbitLeft = document.getElementById('btn-orbit-left');
+export const btnOrbitRight = document.getElementById('btn-orbit-right');
+export const btnZoomIn = document.getElementById('btn-zoom-in');
+export const btnZoomOut = document.getElementById('btn-zoom-out');
+export const placementControlsDiv = document.getElementById('placement-controls');
+export const dpadUp = document.getElementById('dpad-up');
+export const dpadDown = document.getElementById('dpad-down');
+export const dpadLeft = document.getElementById('dpad-left');
+export const dpadRight = document.getElementById('dpad-right');
+export const placementRotateLeft = document.getElementById('placement-rotate-left');
+export const placementRotateRight = document.getElementById('placement-rotate-right');
+export const dpadUpY = document.getElementById('dpad-up-y');
+export const dpadDownY = document.getElementById('dpad-down-y');
+export const placeButton = document.getElementById('btn-place-element');
+export const confirmMoveButton = document.getElementById('btn-confirm-move');
+export const cancelPlacementButton = document.getElementById('btn-cancel-placement');
+export const cancelMoveButton = document.getElementById('btn-cancel-move');
+export const contextMenu = document.getElementById('context-menu');
+export const ctxRotateButton = document.getElementById('ctx-rotate');
+export const ctxMoveButton = document.getElementById('ctx-move');
+export const ctxDuplicateButton = document.getElementById('ctx-duplicate');
+export const ctxDeleteButton = document.getElementById('ctx-delete');
+export const assiseSelect = document.getElementById('assise-select');
+export const createAssiseButton = document.getElementById('btn-create-assise');
