+++ src/elementTypes.js
@@ -1,0 +1,76 @@
+export const elementTypes = {
+    "M50": { name: "Brique M50", dim: [19, 5, 9], color: 0xCC6633 },
+    "M57": { name: "Brique M57", dim: [19, 5.7, 9], color: 0xCC6633 },
+    "M65": { name: "Brique M65", dim: [19, 6.5, 9], color: 0xCC6633 },
+    "M90": { name: "Brique M90", dim: [19, 9, 9], color: 0xCC6633 },
+    "WF": { name: "Brique WF", dim: [21, 5, 10], color: 0xD2691E },
+    "WFD": { name: "Brique WFD", dim: [21, 6.5, 10], color: 0xD2691E },
+    "M50c": { name: "M50 champ", dim: [9, 19, 5], color: 0xB87333 },
+    "M57c": { name: "M57 champ", dim: [9, 19, 5.7], color: 0xB87333 },
+    "M65c": { name: "M65 champ", dim: [9, 19, 6.5], color: 0xB87333 },
+    "M90c": { name: "M90 champ", dim: [9, 19, 9], color: 0xB87333 },
+    "WFc": { name: "WF champ", dim: [10, 21, 5], color: 0xC0C0C0 }, // Example different color
+    "WFDc": { name: "WFD champ", dim: [10, 21, 6.5], color: 0xC0C0C0 },
+    "B9": { name: "Bloc 9", dim: [39, 19, 9], color: 0xAAAAAA },
+    "B14": { name: "Bloc 14", dim: [39, 19, 14], color: 0xAAAAAA },
+    "B19": { name: "Bloc 19", dim: [39, 19, 19], color: 0xAAAAAA },
+    "B29": { name: "Bloc 29", dim: [39, 19, 29], color: 0xAAAAAA },
+    "V1": { name: "Vide 1cm", dim: [40, 19, 1], color: 0xDDDDDD },
+    "V2": { name: "Vide 2cm", dim: [40, 19, 2], color: 0xDDDDDD },
+    "V3": { name: "Vide 3cm", dim: [40, 19, 3], color: 0xDDDDDD },
+    "V4": { name: "Vide 4cm", dim: [40, 19, 4], color: 0xDDDDDD },
+    "V5": { name: "Vide 5cm", dim: [40, 19, 5], color: 0xDDDDDD },
+    "L120_14": { name: "Lint 120x14", dim: [120, 19, 14], color: 0x888888 },
+    "L140_14": { name: "Lint 140x14", dim: [140, 19, 14], color: 0x888888 },
+    "L160_14": { name: "Lint 160x14", dim: [160, 19, 14], color: 0x888888 },
+    "L180_14": { name: "Lint 180x14", dim: [180, 19, 14], color: 0x888888 },
+    "L200_14": { name: "Lint 200x14", dim: [200, 19, 14], color: 0x888888 },
+    "L220_14": { name: "Lint 220x14", dim: [220, 19, 14], color: 0x888888 },
+    "L240_14": { name: "Lint 240x14", dim: [240, 19, 14], color: 0x888888 },
+    "L260_14": { name: "Lint 260x14", dim: [260, 19, 14], color: 0x888888 },
+    "L280_14": { name: "Lint 280x14", dim: [280, 19, 14], color: 0x888888 },
+    "L300_14": { name: "Lint 300x14", dim: [300, 19, 14], color: 0x888888 },
+    "L100_9": { name: "Lint 100x9", dim: [100, 19, 9], color: 0x888888 },
+    "L120_9": { name: "Lint 120x9", dim: [120, 19, 9], color: 0x888888 },
+    "L160_9": { name: "Lint 160x9", dim: [160, 19, 9], color: 0x888888 },
+    "L180_9": { name: "Lint 180x9", dim: [180, 19, 9], color: 0x888888 },
+    "L200_9": { name: "Lint 200x9", dim: [200, 19, 9], color: 0x888888 },
+    "L220_9": { name: "Lint 220x9", dim: [220, 19, 9], color: 0x888888 },
+    "L240_9": { name: "Lint 240x9", dim: [240, 19, 9], color: 0x888888 },
+    "L260_9": { name: "Lint 260x9", dim: [260, 19, 9], color: 0x888888 },
+    "L280_9": { name: "Lint 280x9", dim: [280, 19, 9], color: 0x888888 },
+    "L300_9": { name: "Lint 300x9", dim: [300, 19, 9], color: 0x888888 },
+    "L100_19": { name: "Lint 100x19", dim: [100, 19, 19], color: 0x888888 },
+    "L120_19": { name: "Lint 120x19", dim: [120, 19, 19], color: 0x888888 },
+    "L160_19": { name: "Lint 160x19", dim: [160, 19, 19], color: 0x888888 },
+    "L180_19": { name: "Lint 180x19", dim: [180, 19, 19], color: 0x888888 },
+    "L200_19": { name: "Lint 200x19", dim: [200, 19, 19], color: 0x888888 },
+    "L220_19": { name: "Lint 220x19", dim: [220, 19, 19], color: 0x888888 },
+    "L240_19": { name: "Lint 240x19", dim: [240, 19, 19], color: 0x888888 },
+    "L260_19": { name: "Lint 260x19", dim: [260, 19, 19], color: 0x888888 },
+    "L280_19": { name: "Lint 280x19", dim: [280, 19, 19], color: 0x888888 },
+    "L300_19": { name: "Lint 300x19", dim: [300, 19, 19], color: 0x888888 },
+    "BCA60_9": { name: "BC Ass 60x9", dim: [60, 20, 9], color: 0xE0E0E0 },
+    "BCA60_14": { name: "BC Ass 60x14", dim: [60, 20, 14], color: 0xE0E0E0 },
+    "BCA60_19": { name: "BC Ass 60x19", dim: [60, 20, 19], color: 0xE0E0E0 },
+    "BC60_10": { name: "BC 60x10", dim: [60, 25, 10], color: 0xF0F0F0 },
+    "BC60_15": { name: "BC 60x15", dim: [60, 25, 15], color: 0xF0F0F0 },
+    "BC60_20": { name: "BC 60x20", dim: [60, 25, 20], color: 0xF0F0F0 },
+    "BC60_24": { name: "BC 60x24", dim: [60, 25, 24], color: 0xF0F0F0 },
+    "BC60_30": { name: "BC 60x30", dim: [60, 25, 30], color: 0xF0F0F0 },
+    "BC60_36": { name: "BC 60x36.5", dim: [60, 25, 36.5], color: 0xF0F0F0 },
+    "PUR5": { name: "PUR 5cm", dim: [120, 60, 5], color: 0xFFFF99 },
+    "PUR6": { name: "PUR 6cm", dim: [120, 60, 6], color: 0xFFFF99 },
+    "PUR7": { name: "PUR 7cm", dim: [120, 60, 7], color: 0xFFFF99 },
+    "PUR8": { name: "PUR 8cm", dim: [120, 60, 8], color: 0xFFFF99 },
+    "PUR9": { name: "PUR 9cm", dim: [120, 60, 9], color: 0xFFFF99 },
+    "PUR10": { name: "PUR 10cm", dim: [120, 60, 10], color: 0xFFFF99 },
+    "PUR11": { name: "PUR 11cm", dim: [120, 60, 11], color: 0xFFFF99 },
+    "PUR12": { name: "PUR 12cm", dim: [120, 60, 12], color: 0xFFFF99 },
+    "PUR13": { name: "PUR 13cm", dim: [120, 60, 13], color: 0xFFFF99 },
+    "PUR14": { name: "PUR 14cm", dim: [120, 60, 14], color: 0xFFFF99 },
+    "PUR15": { name: "PUR 15cm", dim: [120, 60, 15], color: 0xFFFF99 },
+    "PUR16": { name: "PUR 16cm", dim: [120, 60, 16], color: 0xFFFF99 },
+    "PUR18": { name: "PUR 18cm", dim: [120, 60, 18], color: 0xFFFF99 },
+    "PROFIL": { name: "Profil", dim: [250, 6.5, 6.5], color: 0xC0C0C0 }
+};
