+++ src/stateVariables.js
@@ -1,0 +1,23 @@
+export let currentTool = 'select'; // Current active tool ('select', 'add', 'move', 'duplicate')
+export let addState = 'idle'; // State for the add tool ('idle', 'positioning')
+export let moveState = 'idle'; // State for the move tool ('idle', 'moving')
+export let selectedElementTypeId = Object.keys(elementTypes)[0]; // ID of the element type selected in the dropdown
+export let jointThicknessCm = parseFloat(jointInput.value) || 1.2; // Thickness of the joint in cm
+export let placedElements = []; // Array to store data of placed elements {id, typeId, mesh, customDimensions?, customName?}
+export let nextElementId = 0; // Counter for assigning unique IDs to placed elements
+export let selectedElement = null; // Reference to the currently selected element's data object
+export let elementForContextMenu = null; // Reference to the element targeted by the context menu
+export let isDragging = false; // Flag to track if the pointer is being dragged (for orbit vs click)
+export let pointerDownTime = 0; // Timestamp when the pointer was pressed down
+export let longPressTimer = null; // Timer for detecting long press
+export let originalMovePosition = null; // Stores the original position when starting a move operation
+export let pointerDownCoords = { x: 0, y: 0 }; // Stores pointer coordinates on pointer down
+
+// Assise Management
+export let assises = [{ id: POSE_LIBRE_ID, name: "Pose Libre", height: null }]; // Array of defined assises {id, name, height}
+export let activeAssiseId = POSE_LIBRE_ID; // ID of the currently active assise
+
+// Three.js Core Components
+export let scene, camera, renderer, controls, raycaster, pointer, axesHelper, gridHelper; // Added gridHelper here
+export let groundPlane, ghostElement; // Ghost element for placement preview
+export const objectsToRaycast = []; // Array of objects the raycaster should check against (ground + placed elements)
